+++ dubbo-samples-cache/src/main/java/org/apache/dubbo/samples/cache/EmbeddedZooKeeper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.samples.cache;
+package org.apache.dubbo.samples.cache;
 
 import java.io.File;
 import java.lang.reflect.Method;
