+++ dubbo-samples-cache/src/main/java/org/apache/dubbo/samples/cache/api/CacheService.java
@@ -17,20 +17,13 @@
  *
  */
 
-package com.alibaba.dubbo.samples.cache;
-
-import org.springframework.context.support.ClassPathXmlApplicationContext;
+package org.apache.dubbo.samples.cache.api;
 
 /**
- * CacheProvider
+ * ValidationService
  */
-public class CacheProvider {
+public interface CacheService {
 
-    public static void main(String[] args) throws Exception {
-        new EmbeddedZooKeeper(2181, false).start();
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/cache-provider.xml"});
-        context.start();
-        System.in.read();
-    }
+    String findCache(String id);
 
 }
