+++ dubbo-samples-cache/src/main/resources/spring/cache-consumer.xml
@@ -28,6 +28,6 @@
 
     <dubbo:registry address="zookeeper://127.0.0.1:2181"/>
 
-    <dubbo:reference id="cacheService" interface="com.alibaba.dubbo.samples.cache.api.CacheService" cache="true"/>
+    <dubbo:reference id="cacheService" interface="org.apache.dubbo.samples.cache.api.CacheService" cache="true"/>
 
 </beans>
