+++ dubbo-samples-cache/src/main/resources/spring/cache-provider.xml
@@ -30,8 +30,8 @@
 
     <dubbo:protocol name="dubbo" port="20880"/>
 
-    <bean id="cacheService" class="com.alibaba.dubbo.samples.cache.impl.CacheServiceImpl"/>
+    <bean id="cacheService" class="org.apache.dubbo.samples.cache.impl.CacheServiceImpl"/>
 
-    <dubbo:service interface="com.alibaba.dubbo.samples.cache.api.CacheService" ref="cacheService"/>
+    <dubbo:service interface="org.apache.dubbo.samples.cache.api.CacheService" ref="cacheService"/>
 
 </beans>
